refactor: redisService 수정

+++ chat-server/src/main/java/DNA_Backend/chat_server/chat/dto/ChatMessage.java
@@ -17,7 +17,7 @@ public class ChatMessage {
     private String roomId;
     private String sender;
     private String message;
-    private int participantCount;
+    private long participantCount;
 
     public void setSender(String sender){
         this.sender = sender;
@@ -27,7 +27,7 @@ public class ChatMessage {
         this.message = message;
     }
 
-    public void setParticipantCount(int participantCount) {
+    public void setParticipantCount(long participantCount) {
         this.participantCount = participantCount;
     }
